+++ wsutil/tempfile.c
@@ -171,7 +171,7 @@ create_tempfile(char **namebuf, const char *pfx)
 {
   static struct _tf {
     char *path;
-    unsigned long len;
+    size_t len;
   } tf[MAX_TEMPFILES];
   static int idx;
 
@@ -218,7 +218,7 @@ create_tempfile(char **namebuf, const char *pfx)
   tmp_file = g_strconcat(tmp_dir, sep, safe_pfx, "_", timestr, "_", TMP_FILE_SUFFIX, NULL);
   g_free(safe_pfx);
   if (strlen(tmp_file) > tf[idx].len) {
-    tf[idx].len = (int)strlen(tmp_file) + 1;
+    tf[idx].len = strlen(tmp_file) + 1;
     tf[idx].path = (char *)g_realloc(tf[idx].path, tf[idx].len);
   }
   g_strlcpy(tf[idx].path, tmp_file, tf[idx].len);
